+++ packages/client/src/Call.ts
@@ -725,6 +725,7 @@ export class Call {
    * @returns a promise which resolves once the call join-flow has finished.
    */
   join = async (data?: JoinCallData): Promise<void> => {
+    const connectStartTime = Date.now();
     await this.setup();
     const callingState = this.state.callingState;
     if ([CallingState.JOINED, CallingState.JOINING].includes(callingState)) {
@@ -793,11 +794,11 @@ export class Call {
           : undefined;
       const { callState, fastReconnectDeadlineSeconds } = await sfuClient.join({
         subscriberSdp: receivingCapabilitiesSdp,
+        publisherSdp: '',
         clientDetails,
         fastReconnect: performingFastReconnect,
         reconnectDetails,
       });
-
       this.fastReconnectDeadlineSeconds = fastReconnectDeadlineSeconds;
       if (callState) {
         this.state.updateFromSfuCallState(
@@ -828,6 +829,16 @@ export class Call {
         clientDetails,
         statsOptions,
         closePreviousInstances: !performingMigration,
+      });
+    }
+
+    // make sure we only track connection timing if we are not calling this method as part of a reconnection flow
+    if (!performingRejoin && !performingFastReconnect && !performingMigration) {
+      this.sfuStatsReporter?.sendTelemetryData({
+        data: {
+          oneofKind: 'connectionTimeSeconds',
+          connectionTimeSeconds: (Date.now() - connectStartTime) / 1000,
+        },
       });
     }
 
@@ -1130,9 +1141,19 @@ export class Call {
    * @internal
    */
   private reconnectFast = async () => {
+    let reconnectStartTime = Date.now();
     this.reconnectStrategy = WebsocketReconnectStrategy.FAST;
     this.state.setCallingState(CallingState.RECONNECTING);
-    return this.join(this.joinCallData);
+    await this.join(this.joinCallData);
+    this.sfuStatsReporter?.sendTelemetryData({
+      data: {
+        oneofKind: 'reconnection',
+        reconnection: {
+          timeSeconds: (Date.now() - reconnectStartTime) / 1000,
+          strategy: WebsocketReconnectStrategy.FAST,
+        },
+      },
+    });
   };
 
   /**
@@ -1140,11 +1161,21 @@ export class Call {
    * @internal
    */
   private reconnectRejoin = async () => {
+    let reconnectStartTime = Date.now();
     this.reconnectStrategy = WebsocketReconnectStrategy.REJOIN;
     this.state.setCallingState(CallingState.RECONNECTING);
     await this.join(this.joinCallData);
     await this.restorePublishedTracks();
     this.restoreSubscribedTracks();
+    this.sfuStatsReporter?.sendTelemetryData({
+      data: {
+        oneofKind: 'reconnection',
+        reconnection: {
+          timeSeconds: (Date.now() - reconnectStartTime) / 1000,
+          strategy: WebsocketReconnectStrategy.REJOIN,
+        },
+      },
+    });
   };
 
   /**
@@ -1152,6 +1183,7 @@ export class Call {
    * @internal
    */
   private reconnectMigrate = async () => {
+    let reconnectStartTime = Date.now();
     const currentSfuClient = this.sfuClient;
     if (!currentSfuClient) {
       throw new Error('Cannot migrate without an active SFU client');
@@ -1196,6 +1228,15 @@ export class Call {
       // and close the previous SFU client, without specifying close code
       currentSfuClient.close();
     }
+    this.sfuStatsReporter?.sendTelemetryData({
+      data: {
+        oneofKind: 'reconnection',
+        reconnection: {
+          timeSeconds: (Date.now() - reconnectStartTime) / 1000,
+          strategy: WebsocketReconnectStrategy.MIGRATE,
+        },
+      },
+    });
   };
 
   /**